refactor(db): move workflow store ownership
